+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Implement `IsEdge` for `HashSet<(usize, usize)>`.
 - Implement `IterVertices` for `HashSet<T>`.
 - Implement `IterWeightedEdges` for `Vec<HashMap<usize, W>>`.
 - Implement `IterWeightedEdges` for `[HashMap<usize, W>; V]`.